Changed IMDB link to open to new tab and added scale for ratings
This changes the IMDB link to open to a new tab instead of replacing the
Streama page.

I also changed the rating on the information page for a movie or episode
to show the scale (for example 3.5/10). I think this is a small, but
good change, because a new user might see a rating lower than 5 and then
think it's out of five. There's also no reason not to show it, because
it doesn't take a lot of space.